pull receiver config into its own prop so we dont have to reinitilize xmtp every time the chat component is rendered. bump version 0.2.5
