+++ package.json
@@ -3,7 +3,7 @@
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
   },
-  "version": "0.2.4",
+  "version": "0.2.5",
   "description": "A simple template for a custom React component library",
   "scripts": {
     "rollup": "rollup -c"
