+++ src/components/ReceiverConfig/ReceiverConfig.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+import {
+  WagmiConfig,
+  configureChains,
+  createClient,
+  defaultChains,
+} from 'wagmi';
+import { alchemyProvider } from 'wagmi/providers/alchemy'
+import { publicProvider } from 'wagmi/providers/public'
+import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet';
+import { InjectedConnector } from 'wagmi/connectors/injected';
+import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
+import { XmtpContextProvider } from '../../xmtp-react/context';
+import { Signer } from 'ethers';
+
+const alchemyKey = 'kmMb00nhQ0SWModX6lJLjXy_pVtiQnjx';
+
+const { chains, provider } = configureChains(defaultChains, [
+  alchemyProvider({ apiKey: alchemyKey }),
+  publicProvider(),
+])
+
+// Set up connectors
+const wagmi = createClient({
+  autoConnect: false,
+  connectors: [
+    new CoinbaseWalletConnector({
+      chains,
+      options: {
+        appName: 'wagmi'
+      },
+    }),
+    new WalletConnectConnector({
+      chains,
+      options: {
+        qrcode: true,
+      },
+    }),
+    new InjectedConnector({
+      chains,
+      options: { shimDisconnect: true },
+    }),
+  ],
+  provider
+});
+
+interface ConfigProps {
+  children: React.ReactNode;
+  signer?: Signer;
+}
+
+const ReceiverConfig = ({ signer, children }: ConfigProps) => {
+
+  return (
+    <WagmiConfig client={wagmi}>
+      <XmtpContextProvider connectedWallet={signer}>
+        { children }
+      </XmtpContextProvider>
+    </WagmiConfig>
+  );
+};
+
+export default ReceiverConfig;
