+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Receiver } from "./Receiver";
+export { default as ReceiverConfig } from "./ReceiverConfig";
